refactor(consensus): #555 P2.1 — split 943 LOC into 9-file package
Codex Round 1 layout (consult: data/llm_consults/2026-05-01_p2-consensus-split.md):

  consensus/
  ├── __init__.py        # facade (263 LOC) — ALL_AGENTS root binding +
  │                      #   analyze_ticker / stream_analyze_ticker /
  │                      #   _compute_weights / agent_readiness /
  │                      #   analyze_portfolio (defined here so monkeypatch
  │                      #   tests work)
  ├── __main__.py        # CLI entry (58)
  ├── models.py          # ConsensusResult, AgentEligibility (48)
  ├── registry.py        # DEFAULT_WEIGHTS, build_all_agents() (53)
  ├── learning_memory.py # _compute_horizon_eligibility,
  │                      #   compute_canonical_weights,
  │                      #   compute_provisional_weights,
  │                      #   select_weight_source (223)
  ├── scoring.py         # _build_consensus pure kernel (160)
  ├── events.py          # _emit_penalty_event_if_fired (58)
  ├── persistence.py     # save_to_recommendations DB upsert (144)
  └── presentation.py    # print_consensus CLI output (132)

All 17 public + private symbols re-exported from package root for back-compat.
Tests that monkeypatch nuri.trading.agents.consensus.ALL_AGENTS still work
because ALL_AGENTS lives at root (build_all_agents() called once in __init__).

Structural separation (#468 codex Round 1) preserved:
  - compute_canonical_weights source contains "outcome_30d" only
  - compute_provisional_weights source contains "outcome_21d" only
  - select_weight_source has no DB I/O
(All three asserted by test_consensus.py::TestStructuralSeparation.)

Verify: 111/111 consensus tests pass. 4461 / 4462 fast suite (1 pre-existing
discord_bot env failure, not from this change). Lint clean. Privacy scan clean.
